refactor: renamed ui components to use dashes
