+++ frontend/cervidae/ui/danger-rating.tsx
